+++ docs/src/changelog.md
@@ -47,6 +47,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 - Rework edition forms as modals:
     - Bookmark tag edition form
     - Feed category edition form
+    - Feed subscription edition form
 
 ## [v0.6.1](https://github.com/virtualtam/sparklemuffin/releases/tag/v0.6.1) - 2026-04-23
 ### Security